requirements.txt
face_recognition
Django==2.1.7
djangorestframework==3.9.2
Pillow==5.4.1
pytz==2018.9
